Remove unused run_demo utility
