+++ api/__init__.py
@@ -1,2 +1,5 @@
 """API utilities for the AutoGen service."""
 
+from .chat_controller import router, create_team
+
+__all__ = ["router", "create_team"]
